# MRG, ENH: Minor speedup of permutation tests