@yh-luo 2020-05-15
### Current results

#### n_perm=100, n_space=100, n_time=100
370 ms ± 4.16 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
15.13% speedup

#### n_perm=1000, n_space=100, n_time=100
3.39 s ± 39.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
11.95% speedup

#### n_perm=1000, n_space=300, n_time=300
5.11 s ± 39.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
12.35% speedup

#### n_perm=1000, n_space=1000, n_time=1000
14.764962434768677
14.28% speedup

The performance is better than I think :partying_face: